devel: Optimize runtime and add evaluation component

+++ src/feature_extractor.py
@@ -4,6 +4,8 @@ import numpy as np
 from PIL import Image
 from scipy import misc
 import pandas as pd
+from scipy import fftpack
+from scipy import signal
 
 from config import *
 
@@ -15,9 +17,39 @@ class FeatureExtractor(object):
         self.img_names = img_names
         self.number_of_snippets = number_of_snippets
         self.accumulated_spectra = np.zeros((SNIPPET_WIDTH, SNIPPET_WIDTH))
-        self.df_header = []
+        # df header
+        self.df_header = [i for i in range(SNIPPET_WIDTH*SNIPPET_WIDTH)]
+        self.df_header = self.df_header  + ['class']
+        # allocate memory
+        #self.fingerprint_df = pd.DataFrame(index=[0], columns=self.df_header, dtype=float)
+        #self.fingerprint_df = pd.DataFrame(columns=self.df_header)
+        self.fingerprint_df = pd.DataFrame()
+
+
+        # for all segments needed for evaluation later
+        max_snippet_amount_per_image = 0
+        #snippets_over_all = 0
+        for image in self.img_names:
+            img_path = self.path_to_class + '/' + image
+            img = misc.imread(img_path)
+            x, y, z = img.shape
+            #snippets_over_all += x//SNIPPET_WIDTH * y//SNIPPET_WIDTH
+            snippet_amount = x//SNIPPET_WIDTH * y//SNIPPET_WIDTH
+            if snippet_amount > max_snippet_amount_per_image:
+                max_snippet_amount_per_image = snippet_amount
+
+        #self.train_data = pd.DataFrame(index=np.arange(0, max_snippet_amount_per_image), columns=self.df_header )
+        self.train_data = pd.DataFrame()
+
+        #check if all are NANs
 
         self.compute_accumulated_spectra(train)
+
+
+        #if snippets_over_all == self.number_of_snippets:
+        #    return("TEST: True. The snippet amount computed in advance is the same than after feature computation.")
+        #else:
+        #    return("TEST: Failed. The snippet amount computed in advance is NOT the same than after feature computation.")
 
 
 
@@ -36,6 +68,8 @@ class FeatureExtractor(object):
 
         segment_count = np.empty([2], dtype=int)
 
+        first_image = True
+
         for image in self.img_names:
             img_path = self.path_to_class + '/' + image
             img = misc.imread(img_path)
@@ -46,7 +80,6 @@ class FeatureExtractor(object):
             segment_count[0] = grey_img.shape[0] // step[0]
             segment_count[1] = grey_img.shape[1] // step[1]
 
-            df_all_segments = self.initalize_train_data(segment_count[0]*segment_count[1])
             sample_no = 0
 
             # for(each segment in img): cut of segment
@@ -55,23 +88,34 @@ class FeatureExtractor(object):
                     # segment as copy of img snippet
                     start_i = i * SNIPPET_WIDTH
                     start_j = j * SNIPPET_WIDTH
-                    segment = grey_img[start_i:(start_i + SNIPPET_WIDTH), start_j:(start_j + SNIPPET_WIDTH)]
+                    segment = grey_img[start_i:(start_i + SNIPPET_WIDTH), start_j:(start_j + SNIPPET_WIDTH), 0]
 
-                    segment_windowed = self.compute_window(segment)
-                    spectrum =self.compute_spectrum(segment_windowed)
+                    # fft only with numpy
+                    #segment_windowed = self.compute_window(segment)
+                    #spectrum =self.compute_spectrum(segment_windowed)
+
+                    window = signal.windows.hann(segment.shape[0])
+                    # fft with scipy
+                    spectrum = fftpack.fft2(segment*window*window.T)
+                    shift = fftpack.fftshift(spectrum)
+                    spectrum = 20 * np.log( np.abs(shift))
 
                     # compute the accumulated fingerprint
-                    np.add(self.accumulated_spectra, np.divide(spectrum, self.number_of_snippets))
+                    self.accumulated_spectra = np.add(self.accumulated_spectra, np.divide(spectrum, self.number_of_snippets))
 
                     self.accumulated_spectra = self.darken_region(self.accumulated_spectra)
 
                     # save sampled spectra for training analysis later
-                    df_all_segments = self.add_train_data(df_all_segments, spectrum, class_name, sample_no)
+                    self.add_train_data(spectrum, class_name, sample_no)
                     sample_no += 1
 
             sample_amount += segment_count[0] * segment_count[1]
 
-            self.save_train_data(df_all_segments, class_name, train)
+            self.save_train_data(self.train_data, class_name, train, first=first_image)
+            self.train_data = pd.DataFrame()
+            #self.train_data.loc[:,:] = np.NaN
+            first_image=False
+            #TODO: test with more images in folder
 
         self.save_fingerprint(self.accumulated_spectra, class_name, sample_amount, train)
 
@@ -128,9 +172,6 @@ class FeatureExtractor(object):
 
 
     def initalize_train_data(self, number_of_rows):
-        # df header
-        df_header = [i for i in range(SNIPPET_WIDTH*SNIPPET_WIDTH)]
-        self.df_header = df_header  + ['class']
 
         # preallocate df
         # create dataframe
@@ -140,20 +181,29 @@ class FeatureExtractor(object):
         return(df)
 
 
-    def add_train_data(self, train_data, spectrum, class_name, sample_no):
+    def add_train_data(self, spectrum, class_name, sample_no):
 
         # add segment fft to data_detailed
-        new_sample = spectrum.reshape(spectrum.size).tolist() + [class_name]
+        new_sample = np.array(spectrum.reshape(spectrum.size).tolist())
+        new_sample = new_sample.reshape(spectrum.size, 1)
+        new_df = pd.DataFrame(new_sample.T, columns=self.df_header[:-1])
+        new_df['class'] = class_name
 
         #df.append(new_sample, ignore_index = True)
 
-        train_data.loc[sample_no] = new_sample
+        #self.train_data.loc[sample_no] = new_sample
+        if self.train_data.empty:
+            self.train_data = new_df
+        else:
+            #self.train_data.loc[len(self.train_data)] = new_sample
+            self.train_data = self.train_data.append(new_df, ignore_index=True)
 
-        return(train_data)
+
+        return()
 
 
 
-    def save_train_data(self, df_all_segments, class_name, train):
+    def save_train_data(self, df_all_segments, class_name, train, first=False):
 
         if train:
             folder = SUBPATH + '/trained_samples/'
@@ -165,9 +215,9 @@ class FeatureExtractor(object):
         if not os.path.exists(folder):
             os.makedirs(folder)
 
-        if os.path.isfile(file_name):
+        if os.path.isfile(file_name) and not first:
             unpickled_df = pd.read_pickle(file_name)
-            df_all_segments = unpickled_df.append(df_all_segments, ignore_index=True)
+            df_all_segments = unpickled_df.append(df_all_segments.dropna(axis=0), ignore_index=True)
             df_all_segments.to_pickle(file_name)
         else:
             df_all_segments.to_pickle(file_name)
@@ -196,9 +246,13 @@ class FeatureExtractor(object):
         magnitude_all = magnitude_all - magnitude_all.min()
         magnitude_all = magnitude_all / magnitude_all.max() #sum()
         #TODO: check what was meant with sum()?
+        # self.fingerprint_df.fillna() or something to refill all values with NANs
+        #self.fingerprint_df.loc[0] = magnitude_all.reshape(magnitude_all.size).tolist() + [class_name]
+        #self.fingerprint_df.append(magnitude_all.reshape(magnitude_all.size).tolist() + [class_name])
+        feature_vector = np.array(magnitude_all.reshape(magnitude_all.size).tolist() + [class_name])
+        feature_vector = feature_vector.reshape(magnitude_all.size+1, 1)
+        self.fingerprint_df = pd.DataFrame(feature_vector.T, columns=self.df_header)
 
-        fingerprint_df = pd.DataFrame(index=[0], columns=self.df_header)
-        fingerprint_df.loc[0] = magnitude_all.reshape(magnitude_all.size).tolist() + [class_name]
 
         # save image: merged frequency spectrum by addition
         f_add = Image.fromarray(np.divide(magnitude_all, magnitude_all.max()) * 255).convert('RGB')
